test: Switch to InSpec verifier
