+++ .kitchen.yml
@@ -5,6 +5,9 @@ driver:
 provisioner:
   name: chef_zero
 
+verifier:
+  name: inspec
+
 platforms:
   - name: ubuntu-16.04
     named_run_list: debian
@@ -52,8 +55,7 @@ suites:
           acl_datacenter: FortMeade
           acl_default_policy: deny
   - name: client
-    provisioner:
-      policyfile: test/fixtures/policies/default.rb
+    named_run_list: client
     attributes:
       consul:
         config:
