+++ TODO.md
@@ -1,3 +1,2 @@
 # TODO
 - Fix the helpers and clean up the where it injects into DSL.
-- Use the InSpec instead of ServerSpec.
