+++ test/fixtures/policies/default.rb
@@ -7,6 +7,7 @@ named_run_list :centos, 'sudo::default', run_list
 named_run_list :debian, 'apt::default', run_list
 named_run_list :freebsd, 'freebsd::default', 'sudo::default', run_list
 named_run_list :windows, 'windows::default', run_list
+named_run_list :client, 'consul::default'
 
 default['authorization']['sudo']['users'] = %w(kitchen vagrant)
 default['authorization']['sudo']['passwordless'] = true
