+++ test/integration/spec_helper.rb
@@ -1,0 +1,7 @@
+def windows?
+  (/cygwin|mswin|mingw|bccwin|wince|emx/ =~ RUBY_PLATFORM) != nil
+end
+
+def consul_version
+  '0.7.2'
+end
