+++ test/integration/webui/webui_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require_relative '../spec_helper'
 
 consul_version = '0.7.1'
